services/microcred/pyproject.toml
[tool.poetry]
name = "microcred-genai"
version = "0.1.0"
description = "Monorepo skeleton for the MicroCred credentialing platform."
authors = ["MicroCred Team <dev@microcred.local>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
fastapi = "0.110.2"
pydantic = "2.8.2"
sqlalchemy = { version = "2.0.25", extras = ["asyncio"] }
structlog = "24.1.0"
rq = "1.15.1"
redis = "5.0.1"
pyjwt = { version = "2.9.0", extras = ["crypto"] }
passlib = { version = "1.7.4", extras = ["bcrypt"] }
python-dotenv = "1.0.1"
uvicorn = { version = "0.30.1", extras = ["standard"] }
slowapi = "0.1.9"
pydantic-settings = "2.4.0"
alembic = "1.13.2"
asyncpg = "0.29.0"
aiosqlite = "0.20.0"
psycopg2-binary = "2.9.9"
email-validator = "^2.3.0"
bcrypt = "4.0.1"
requests = "^2.32.5"

[tool.poetry.group.dev.dependencies]
pytest = "8.2.1"
pytest-asyncio = "0.23.6"
pytest-mock = "3.14.0"
httpx = "0.27.0"
ruff = "0.5.7"
mypy = "1.10.0"
pre-commit = "^4.5.1"

[build-system]
requires = ["poetry-core>=1.8.2"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
src = ["src"]
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["B008"]  # Allow function calls in argument defaults (needed for FastAPI Depends)

[tool.ruff.format]
preview = true

[tool.mypy]
python_version = "3.11"
exclude = ["tests/", "alembic/", "scripts/"]
explicit_package_bases = true
